services/subscription-service/requirements.txt
flask
pymongo
requests
pyjwt
